requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
psycopg2-binary==2.9.9
pgvector==0.2.5
python-dotenv==1.0.1
unstructured==0.14.6
pypdf==4.3.1
psycopg[binary,pool]


# --- HF embeddings stack ---

# --- HF embeddings stack (pinned for stability) ---
sentence-transformers==2.7.0
transformers>=4.41,<5
torch>=2.1.0

# --- HF embeddings stack (managed by hf_apply_patch_and_reset_v2) ---
sentence-transformers==2.7.0
transformers>=4.41,<5
torch>=2.1.0
# --- /HF embeddings stack ---

